backend/requirements.txt
# Phase 1 Dependencies
# Install with: pip install -r requirements.txt

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client - pinned for Windows ARM
httpx==0.25.1

# SSE support
sse-starlette==1.8.2

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0

# Config and validation - pinned for stability
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Galileo AI Observability
galileo[openai]>=0.1.0

